Update patch set 1
Patch Set 1:

> I like it, though I have the feeling Nicolas won't.

Thanks!  Yeah, this is just a proposal.  We had a quick chat with David and Calin about this earlier.  I should probably write a short document to explain the pros and cons of this approach (and other strategies as well).

Patch-set: 1
Reviewer: Roland Levillain <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0