Update patch set 12
Patch Set 12:

> I see why you added the arch to each line and not just the START
 > but how about this...
 > 
 > In _extractLine, try to parse all the CHECK-START(-ARCH)
 > combinations and store the arch that matched with the TestCase
 > object. Then you just need to change the RunTest logic to skip
 > TestCases which don't match the target arch specified on the
 > command line and you'll have avoided the parser issue.

If you prefer I'll do that. But since __extractLine does not have access to the TestCase object that involves modifying the prototype of fnProcessLine to return a triplet, and so modifying the code for the c1visualizer.
What about returning an object with properties line, lineNo, arch instead?

Patch-set: 12