Update patch set 2
Patch Set 2:

(1 comment)

> > don't you do that just because DCE gets more precise during
 > > RemoveDeadInstructions()?
 > 
 > Not really. It's because RemoveDeadBlocks does not update Phis, so
 > they might be left with only one input (say when you remove one
 > if-branch). Running phi elimination avoids duplicating code.

... we can discuss that on another thread since it's unrelated (but if RemoveDeadInstructions does not benefit of SsaRedundantPhiElimination, then why are you doing it there?)

 > My point being, if some pass is an inseparable part of another
 > optimization, why not let it manage it itself? What you're doing
 > here is extracting logic from the inliner (and making a mess in the
 > process).

the extra passes we run are not "inseparable" from the inliner. They merely benefit of the inliner. If you want an example of dependencies you should look at side-effects -> lcim. We still do GVN in between because that's better for lcim. Under your proposal, achieving a similar effect would be impossible.

I can extract this into a function let's say OptimizingCompiler::Inline() if you think that adds clarity. This is what we also do with AllocateRegisters (another example of real dependencies which we keep separate).

Patch-set: 2