Update patch set 1
Patch Set 1:

Some comment on the general algorithm:
Since we've already done loop analysis (essentially finding SCC's, plus more like depth of loop. Also we only deal with natural loops), it looks to me it can be simplified on iv detection side with the loop info available.

Each iv must be a loop phi, or derived from another iv. To check whether a loop phi is an iv, the input of the loop phi is traced all the way back to itself, with only certain ops allowed. It's like detecting SCC's but can be simpler to understand/implement. No need for full Tarjan's algorithm. Does this do the same thing?

Patch-set: 1
Label: Code-Review=0