Update patch set 1
Patch Set 1:

(5 comments)

My knowledge is a bit old, but in my days using rep xxx was not considered the most optimal way of doing this. Have you considered using the pcmp[b,w,d] family (with a pre-loop to obtain 16-byte alignment; and a post-loop to avoid overshooting the compared data if necessary). Of course, this will require the xmm registers, not sure if we have that support already.

Patch-set: 1