Update patch set 6
Patch Set 6:

> (2 comments)
 > 
 > Still no luck with test case yet. Will check with shertz.

Maybe it's because I tested with Quick (but it requires to update the JDWP makefile). Could you upload a CL disabling the baseline compiler so I can test with the Optimizing compiler? In the meantime, I'll disable it locally (but the last time I tried, it crashed the compiler...)

Patch-set: 6