Update patch set 5
Patch Set 5: Code-Review-1

As long as we're still compiling with GCC, it's the way to go. So please don't cut out thumb_func, considering that both ART and assembly stubs are mixed-mode (and that's what's supposedly special about thumb_func).

I'll patch your patch (had one in the making, anyways).

Also, again -1 to switching over the integrated assembler in the same CL. That is a separate concern.

Patch-set: 5
Label: Code-Review=-1