Update patch set 3
Patch Set 3:

Sorry for the late comment. I was also on holiday and didn't manage to take a look before leaving. 

Based on the comments in the document I was left with the impression that most of us prefer the plain old way?

Patch-set: 3
Reviewer: Calin Juravle <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0