Update patch set 6
Patch Set 6:

> > > This is equivalent to my 'backend optimization' pass, so I am
 > > okay
 > > > with it, but I prefer a scheme that removes the machine
 > dependent
 > > > optimizations out of optimizing_compiler.cc and makes it BE
 > > > specific.
 > >
 > > So something like codegen_->RunOptimizations? I think that could
 > > work for me. WDYT Alexandre?
 > 
 > That sounds good. Initially I thought there would be multiple
 > arch-independent passes like `InstructionSimplifierArch` that would
 > then dispatch to specific arch visitors. Now that this wrapper has
 > been removed, and considering the arch-specific opts will probably
 > diverge at the beginning I think that is a good idea. I'll update
 > the code.

Now I'm not sure.

In Mark's patch extracting the code out of optimizing_compiler.cc is easy because it is a single pass.
Moving the run of arch-specific passes to code_generator_<arch>.cc involves
- Sharing code that is currently assumed to only be used in optimizing_compiler.cc (`RunOptimizations()`, `class PassObserver`, class `PassScope`). I have extracted that into a new header `optimizing_compiler_helper.h`. I'm afraid this would raise cherry-picking issues.
- Making `code_generator_<arch>.cc` include headers for the passes used (here gvn.h, side_effects_analysis.h, instruction_simplifier_arm64.h). I don't think this really belongs in the code_generator<arch>.cc files.

I prefer the current state of the code. Yes there is a switch for the different archs, but the only thing there is a list of passes so it should not be invasive. It also has the advantage of gathering in one place the passes run by the compiler (arch-independent and arch-specific).

Opinions?

Patch-set: 6