Update patch set 3
Patch Set 3: Code-Review-1

(8 comments)

The new interpretation of --patched-image-file is only meaningful for patching oat files, not image files, right?

I think it would be cleaner to have a single patch function that accepts an optional image and an optional oat file to patch, instead of having three separate patch functions, one for image only, one for image and oat file, and one for oat file. This way you wouldn't need to duplicate the code, and you could put the new code for determining the patch delta with the existing logic that calculates the patch delta from original and new base offsets.

Patch-set: 3
Label: Code-Review=-1