Update patch set 10
Patch Set 10:

> I'm here if you need me to fix things!

Thanks Guillaume. I will have time to get back to this next week. I want to run some benchmarks first, and think were is the best place to actually make the graph transformation. Historically we pushed this to the simplifier, but maybe it's getting too much now.

Patch-set: 10