Update patch set 12
Patch Set 12:

> >
 > > One drawback of interleaving arch-independent and arch-specific
 > > passes may be the checker tests. This can become a mess if we try
 > > to check a pass occurring after an arch-pass. Having a more
 > > flexible way of specifying what passes should run would help
 > there.
 > 
 > I think we should be fine as long as each pass has a unique name,
 > right?

Not really. The arch-specific passes may change how the graph looks like.

Example:

Passes run:
    pass1
    instruction_simplifier_arch
    pass2

    CHECK-START: pass1
    ...
    CHECK:       ArrayGet
    CHECK:       Add [[Arg]] [[Const0]]

    CHECK-START: pass2
    ...
    CHECK:       ArrayGet
    CHECK-NOT:   Add

This would fail on arm64 because the ArrayGet was split. We can work around that by doing arch-specific checker tests. But we ideally don't want to resort to arch-specific tests when the arch-specific parts are not what we want to check.

Patch-set: 12