Update patch set 1
Patch Set 1:

Will try to do, however I guess it will not be a Java test because the bug heavily depends on behavior of register allocator. I hope to create a test tomorrow.

Please let me know if you come up with some solution. I've tried to optimize this method (in terms to avoid unnecessary spills) but it is not completed yet.

Patch-set: 1