Update patch set 1
Restored

I believe your test should work for my case as well. In reality this change is not enough for efficiency. To make it more efficient we need to iterate from end to the beginning knowing the fact that split pairs on x86 mostly come from hardcoded eax:ecx.

Patch-set: 1
Status: new