Update patch set 2
Patch Set 2:

At this time, rewriting the buffer seems like a lot of byte movement.  We also have to handle variable sized instructions.  We can always implement it later.  ShortLabels can either be removed at that point, or just left in, as they are known not to need rewriting.

Patch-set: 2
Label: Verified=0