Update patch set 13
Patch Set 13:

I've got read barrier tests running on N9 right now but I don't want to interrupt that to check if I also turned on the heap poisoning. However, the host tests in this configuration have already given me a failure in 466-get-live-vreg64:

+art F  4782  4782 art/test/466-get-live-vreg/get_live_vreg_jni.cc:46] Check failed: GetVReg(m, 0, kIntVReg, &value) == false (GetVReg(m, 0, kIntVReg, &value)=1, false=0) 

When I finish this test run I'll check if it's a pre-existing failure. I don't have a clue how this CL could be responsible.

Patch-set: 13