Update patch set 1
Patch Set 1:

> > > > > Do we want to keep this?
 > > > > >
 > > > > > I somehow feel that we shouldn't really support debug
 > symbols
 > > > for
 > > > > > the JIT. If people want to do native/performance debugging,
 > > use
 > > > > an
 > > > > > AoT build.
 > > > > >
 > > > > > So I'd vote for removing the gdb jit hook completely,
 > > > considering
 > > > > > it was never adapted to the mcjit (it triggers on opening
 > an
 > > > ELF
 > > > > > file).
 > > > >
 > > > > What's ICF.
 > > >
 > > > Identical Code Folding. Same as our dedupe. Problem is that the
 > > > current implementation (it's intentionally empty) gets merged
 > > into
 > > > other things, and GDB registers a call even when we don't want
 > to
 > > > register code...
 > > >
 > > > >
 > > > > Getting the gdb to know about JIT frames still helps us (ART
 > > > > developers) when debugging.
 > > >
 > > > Well, then there's a major revamp necessary. None of this works
 > > > right now, and most of the infrastructure is missing.
 > >
 > > What do you mean? I understood that calling the gdb hook was
 > enough
 > > to map native pc -> name.
 > 
 > You need to give it a pointer to an eh_frame right now, at least
 > that's the path I can see.

Based on my understanding we have to produce a valid ELF file in memory and then specify it in the JIT list. If it have a symtab, then we will have symbol names, if it have more debug info, then the debugger can display other things too (e.g. line information)

eh_frame is required if we want gdb to show a stack trace, but if we use lldb, then usually it isn't necessary (assuming we have symtab) for almost anything

Patch-set: 1
Label: Verified=0