Update patch set 2
Patch Set 2:

(4 comments)

A clear win from this change is the use of the "for (a : b)" syntax which feels a lot nicer than what we have today. A clear downside is the IMO broken stl API. Personally that downside is a pretty big one. Beside the 'for' syntax, what are the advantages of moving to this API?

Probably everyone involved in that code base should chime in.

Patch-set: 2