Update patch set 6
Patch Set 6:

Hi Yevgeny, thanks for you patience.

I would like to comment on the mandatory/optional part.  I think I know what your aim is since that was my first instinct as all - accurately annotate all original instructions and pick any reasonable nearby dex_pc for the synthesized instructions.  After all, some mapping is better then none, and we should aim for 100% coverage.

I have since changed my opinion.  Guessing information can be more harmful than not having the information in the first place.  Imagine the following two source lines:
 a = b + 1
 c = d + 1
Which might, in theory, be compiled to something like:
 mov r0, #1  // dex_pc = 2
 add r1, r2, r0  // dex_pc = 1
 add r3, r4, r0  // dex_pc = 2
All of the instructions are mapped to something, but native instructions for given dex_pc are not continuous.  I suppose that is ok for profiles, but it makes life of a debugger very difficult.  If you want to stop at the start of the second line, when do you stop?

I would rather have:
 mov r0, #1  // dex_pc = kNoDexPc
 add r1, r2, r0  // dex_pc = 1
 add r3, r4, r0  // dex_pc = 2
Followed by some post-processing to assign some dex_pc to all instructions.  For example, you can visit all HInstuctions in post processing pass, and if it has kNoDexPc, assign it the dex_pc of the instruction which uses it.  Or assign it the dex_pc of the previous/following instruction.

The point is, the post-processing heuristic might actually be in better position to guess reasonable dex_pc based on location of the instruction and its use (shared code vs unique).

So I would prefer to use kNoDexPc liberally for synthesized instructions, and add some back-end heuristic to set it to something reasonable.  As a bonus it keeps the patch small.

Patch-set: 6
Reviewer: David Srbecky <1065473@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0