Update patch set 5
Patch Set 5:

Nicolas, please take a look at the sample arm implementation. Is that OK? (If so, I'll complete it).

The downside to emitting the invoke-virtual code in the slow-path is that it won't be a direct-pointer call, even if possible. But I don't think that's necessary, considering it's a slow-path...

Patch-set: 5
Label: Verified=0