Update patch set 7
Patch Set 7:

(11 comments)

> (2 comments)
 > 
 > Is there ever a time we can make use of a corrupt or invalid image
 > file?
We sometimes want to look at images which would not be selected by the runtime or know more about the place the image is, even if the file there is invalid.

 > Could we treat it the same as a missing image file (with some
 > vlogging in the image assistant to note the distinction if
 > desired)?
We already do.

 > Can you summarize all of the relevant image states with
 > two booleans:
 > 1. We have or don't have an uncorrupted system image.
 > 2. We have or don't have an uncorrupted cached image.
 > 
 > That would be much easier for me to understand than the current
 > ImageState + ImageInfo combination.

No We need the more expansive states to do image_space.cc

Patch-set: 7
Label: Verified=0