Update patch set 3
Patch Set 3:

> (4 comments)
 > 
 > This looks like a lot of change/infrastructure for very minor
 > benefits. Wouldn't it be simpler to just consider runtime calls to
 > always GC?

I'll look at the comments on the code a bit later.

For now indeed it is not very important. However Ningsheng is working on HInstruction scheduling, and I believe it will be relevant in this area. (The scheduling work is not at a stage where we can confirm this yet.)

Another way of achieving something good enough for scheduling would be to only manually pass a 'GC safe' parameter to `InvokeRuntime` defaulting to `false`. This way we may be able to avoid a lot of the infrastructure changes.

Patch-set: 3
Reviewer: Alexandre Rames <1052304@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0