Update patch set 1
Patch Set 1:

The compiler was warning about the function missing the noreturn attribute, because the UNREACHABLE() block was being hit "unconditionally" from its point of view, e.g.

   constexpr int kUseReadBarrier = true;
   if (kUseReadBarrier) { LOG(FATAL); UNREACHABLE(); }

IMHO I think this might be a compiler warning bug (overly aggressive) because being behind a "if(CONST_EXPR) { UNREACHABLE();" should not trigger this warning because if CONST_EXPR is ever false then why would you put the attribute there? It doesn't make sense. If it's with macros the noreturn could be conditional, but we can't make conditional noreturns with constexpr.

It's not the first time I've ran into overly aggressive warnings with constexpr/templates.

Patch-set: 1