Update patch set 1
Patch Set 1:

@Andreas:
Not sure how to test this, as unaligned access will be handled by either hardware or kernel. Not having unaligned accesses is just faster, but result is the same. So, even if we have regressions, they would only affect performance.

Only idea on my mind is to generate Java code that 'might' have unaligned access, and then check that generated mips64 code is *exactly* what we expect it to be. Do you have a similar test we can use as a reference? Also, maintaining such test would be difficult.

Patch-set: 1