Update patch set 1
Patch Set 1:

> @Andreas:
 > Not sure how to test this, as unaligned access will be handled by
 > either hardware or kernel. Not having unaligned accesses is just
 > faster, but result is the same. So, even if we have regressions,
 > they would only affect performance.
 > 
 > Only idea on my mind is to generate Java code that 'might' have
 > unaligned access, and then check that generated mips64 code is
 > *exactly* what we expect it to be. Do you have a similar test we
 > can use as a reference? Also, maintaining such test would be
 > difficult.

It's not that difficult to create a test case. You just need a method with enough parameters that the last arg is passed on the stack and make sure it's wide and held in a register pair starting with an odd-numbered register (use smali). The method than needs to use the parameter to exercise the load and you must somehow prevent it from being inlined, so that the caller can exercise the store.

You could benchmark the invoke against a similar method with aligned loads to make sure we're not using the expensive workaround in kernel. (I asssume handling in hardware should be fast enough for us not to care that much.)

Patch-set: 1
Label: Code-Review=0