Update patch set 2
Patch Set 2: Code-Review-2

After discussion with Dan & Stephen, we may need to reconsider. Gold is also used for x86 and x86-64, and we *may* switch arm64 at some point.

According to Dan we use --icf=save, so an alternative would be to take the address of the function somewhere, so that gold would not be allowed to collapse.

Patch-set: 2
Label: Code-Review=-2