Update patch set 2
Patch Set 2:

Gold is currently used on x86/x86_64/arm but ICF is only enabled on arm (https://android.googlesource.com/platform/build/+/316f720/core/combo/TARGET_linux-arm.mk)

I think the best approach would be to detect when ICF is enabled with setting a global variable in TARGET_linux-arm.mk and then set --icf=safe and --keep-unique based on this but we have to discuss it with the android build team.

Taking the address of the symbol (and storing it in an exported variable to avoid some optimizations) wouldn't necessarily help because it is still a valid optimization to merge a function where the address is taken with an other one where the address isn't taken.

I am fine with the solution what Oleksiy suggested (adding VLOG) as it will fix the issue in every scenario but it is a little bit of hack to change the function to disable an optimization.

Patch-set: 2
Label: Verified=0