Update patch set 1
Patch Set 1:

> > > Not sure I like this "fix," it's fixing a symptom...
 > >
 > > I somewhat agree but OTOH if we chose the delta of 0 we can
 > > actually run without relocation enabled and so should. In that
 > case
 > > we would expect that we can execute.
 > 
 > So if you're the lucky guy that gets a 0 delta, you won't pay for a
 > duplicate boot image? Randomness for the win. I'd go with
 > consistency and always relocate.

No. 

This is only a problem in run-tests since we place the oat file in ANDROID_DATA already compiled. If we picked 0 as the delta we see it in ANDROID_DATA and assume it must have been compiled against the relocated image or already relocated since it is in ANDROID_DATA and matches the delta & checksum.

Making this work in run-tests would require rewriting the run-test script quite a bit.

Patch-set: 1