Update patch set 2
Patch Set 2:

(2 comments)

Thank you for the test, Mark. Please write more though. The method `p` only tests that the PackedSwitch gets removed correctly if it is inside a dead if branch. It would be good to test the changes to DCE's MarkReachableBlocks as well, both when the constant switch value matches a case branch and when it doesn't.

Patch-set: 2
Label: Verified=0