Update patch set 3
Patch Set 3:

Hello Mark, 
Thanks for the CL, this has been on my TODO list for a while. However, I suggest you make HSelect an arch-independent instruction - it is a useful HIR abstraction and the GenerateSelects pass would become a generalization of BooleanSimplifier (which I wrote for boolean negation). You could still consult codegen on when the simplification is actually worthwhile. What do you think?

Also, please write tests for this. If needed, you can tell Checker to run a test only on a given architecture (check out test 526 for examples). It didn't work on host until recently but is fixed now.

Patch-set: 3
Label: Verified=0