Update patch set 2
Patch Set 2:

> wouldn't this get in the way when I test locally with the buildbot
 > config? i.e. i'll have to set up the linker myself.

Unfortunately yes. But it looks like ninja will mess up with you anyway: since we use make to run all tests, if ninja sees a change in the environment, it will recompile stuff.

For you, it's not too difficult to move your phone to an updated version. For bots it's a different problem.

Patch-set: 2