Update patch set 7
Patch Set 7:

I have added support for using booleans as the input to the HIf that generates a HConditionalSelect.  Previously, it only handled HCondition as an input.

While I agree that merging using the condition code for multiple selects is a great idea, I am leery of doing this.  On the X86, the condition code is a fickle thing, and needs to be calculated right before the use.  Perhaps ARM can detect this case just before code generation and generate better code, or check that the previous instruction is a HConditionalSelect with the same inputs?

As to extending the live range, it does keep the inputs to the HCondition live longer.  This is one more value live than there would be without the HConditionalSelect, as the value of the HCondition would need to remain live in any case.

Perhaps this should be a per-arch decision?

Another question:  Are FP compares a good candidate for this?  I think that the Dex bias makes it hard to generate a single condition code (at least on X86), but I can add support for the Bias in HConditionalSelect if it is useful to other architectures.

Patch-set: 7
Reviewer: Mark P Mendell <1036869@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0