Update patch set 2
Patch Set 2:

hmm, I think you misunderstood me, no real problem here :). This is why I already +2 it.

I thought the script is designed for both, our use, and the buildbot use. It's much more convenient when I don't have to keep track what env to set. I asked for a comment just to not forget about it, but it's fine without, especially if this is mainly for buildbots.

Patch-set: 2