Update patch set 2
Patch Set 2:

This is still work in progress, but please have a look and let me know what you think.

It changes the BooleanSimplifier to generate a generic HSelect (3 inputs: 2 values + condition) later simplified by InstructionSimplifier. The fallback implementation on all platforms is if-then-else but CMOV/CSEL is used if possible.

There's a couple of things missing:
(1) We should try to move the HCondition before the HSelect.
(2) BooleanSimplifier could extract an instruction with no side effects from the branches, like Mark did in his CL.

Patch-set: 2