Update patch set 9
Patch Set 9:

(2 comments)

We enable lock counting unconditionally all the time (as opposed to certain verification runs) but limited to methods that failed the verification and the reason that we don't need to do this in the compiler is that we won't compile such methods to begin with. Correct?

Patch-set: 9