Update patch set 2
Patch Set 2:

> (5 comments)
 > 
 > Not sure what is the goal of this change, can you please explain?

(Sorry for the late reply)

We have not been using the Interpreter->Interpreter or Interpreter->CompiledCode functions as TLS-stored entrypoints in a long time. I believe it was ngeoffray who did the final rewrite of this. This CL is pure cleanup:

 Free up TLS slots (saves 8B/16B)
 Remove "extern" definitions
 Move code close to callers where possible

As Vladimir mentioned, the function you have style issues with was pure copy-n-paste (with slight renaming). You are right that it doesn't comply to style anymore, but for such refactorings I prefer a clean-up CL, as it makes tracking changes easier. I'll put the cleanup on my TODO list and hopefully get it done today or tomorrow.

Patch-set: 2