Update patch set 1
Patch Set 1:

> Didn't we say it wasn't worth special casing? What made you change
 > your mind?

You suggested it wasn't worth it - I'm a perfectionist. I think the overhead is minimal, and it's rather clean. (I do agree that we should figure out why we see code locking on null, but that's a different issue.)

With this change, only tts is left with locking failures.

Patch-set: 1