Update patch set 1
Patch Set 1:

I'd like to see a better mix of benchmarks, not just large apps.

I'd like to see the memory overhead you add (arena things don't go away immediately, you likely need to change how the fallthrough line works - it should be trivially shareable if it's allocated in CodeFlowVerifyMethod; the new class->regtype map).

I'd like to see which of the modifications actually is beneficial. My guess is that your class->regtype map actual is responsible for a majority of the speedup.

Patch-set: 1