Update patch set 1
Patch Set 1:

(12 comments)

> It's not functional yet, somehow the base label ends up lower than
 > the buffer size when emitting.

You have missed the fact that CodeGeneratorARM::Finalize() first calls __ FinalizeCode() and then adjusts the block labels.

I would pull the CreateJumpTable() out of the assembler and just ask the assembler to create a Literal for the jump table, allowing a zero-literal of arbitrary size (not bound by the kMaxSize = 8). The code generator would then have to fill in the table after adjusting the block labels.

(Alternatively, we could push all the labels we need adjusted to the assembler and have CodeGeneratorARM::Finalize() adjust only non-Label information.)

Patch-set: 1
Label: Code-Review=0