Update patch set 2
Patch Set 2: Code-Review+1

This change is required for a patch to runtime/thread.h that we expect to submit in a few days, and has been verified to fix the problem there.

This script is using regexps to parse a context-free grammar (C++) which is by definition impossible to do correctly in every case. This patch improves the situation, but the script will still fail if, for example, someone wrote a comment after the '}' and before the 'while' or ';'.

Patch-set: 2
Reviewer: Bruce Hoult <1085148@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1