Update patch set 2
Patch Set 2:

Vladimir, the issue is now that the anchor label actually isn't fixed up. (A simple test that doesn't move the anchor location actually works)

Would you be OK with this implementation - still keeping the details in the assembler? If so, should I just add an empty Fixup type so that the label will be shifted correctly?

Patch-set: 2
Label: Verified=0