Update patch set 1
Patch Set 1:

> Should we generalize this a bit?
 > 0>unsigned -> false
 > 0<=unsigned -> true
 > unsigned<0 -> false
 > unsigned>=0 -> true

Sure. I was already tempted to do that, but looking around I was not sure if we want to safe time/space by just doing typical cases, or be rigorous and optimize all. Glad to see you like the latter also :-)

Also, do you have good ideas to unit test this. I really like the constant_folding_test infrastructure, but since this is not a bytecode instruction, I have to set up the HIR very differently.

Patch-set: 1