Update patch set 1
Patch Set 1:

(1 comment)

The main motivation is the disassembly will be sorted according to the address. And it will be easier for a stupid script to analysis it. For example, if we want to search for some disassembly code pattern which mades up by instructions in different blocks.

Patch-set: 1
Label: Verified=0