Update patch set 1
Patch Set 1:

> > String.equals is the parent to this change because in his
 > comments
 > > regarding CL 179160 Andreas Gampe advised "[p]lease consider
 > > String.equals as one of the first intrinsics to implement...." so
 > I
 > > submitted that CL 179971 before submitting this change. Whatever
 > > order the Google team agrees to, I'm happy to conform to. Just
 > let
 > > me know what is decided.
 > >
 > > Thanks.
 > 
 > Sure, but code-wise, there is nothing dependent between the two,
 > isn't it?

Except that your Verified + 1 isn't worth anything, as this code won't run at the moment. The cutout for Mips is still in - no intrinsics enabled.

This was required for the NeedsEnvironmentOrCache change to allow inlining StringEquals, and as long as StringEquals isn't implemented, we turn off *all* intrinsics (we probably should make this granular, then this CL could be verified and go in).

That's the reason I said String.equals should be accelerated...

Patch-set: 1
Label: Code-Review=0