Update patch set 2
Patch Set 2:

(6 comments)

Thanks for quick review.

| Not sure the Proxy -> AnyProxy rename is worth doing.

That was to catch every single call site of it and change it to either AnyProxy, ReflectProxy, or to handle both cases. I couldn't just blindly let the Is"Proxy" check do both (return true if its a lambda OR a reflect proxy) because a handful of places needed to make a distinction.

I also didn't want to just do "IsProxy || IsLambdaProxy" in a ton of places which also resulted in a huge change that made it more brittle in the future.

Patch-set: 2