Update patch set 8
Patch Set 8:

> I would love to help you out.  I don't know MIPS code at all (even
 > less than ARM!).  Do you have an questions that I can help with? Do
 > you want to run some possible code by me?
 > 
 > The basic change is to check if the jump can be reversed, and if
 > so, jump on the opposite condition to the false target.

We can figure it out and implement the equivalent for MIPS as well, it's just that the change is rather generic and straightforward and it would've been nice to have it implemented for most architectures as it's desirable for all of them (I've seen this jump issue in MIPS disassembly sometime ago). No worries.

Patch-set: 8
Reviewer: Alexey Frunze <1057043@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0