Update patch set 4
Patch Set 4:

> (5 comments)
 > 
 > Comments so far. Could we talk at the meeting about it? It's a bit
 > hard to follow the problem with the code.

Ok, will talk at the meeting. Basically it's whenever a merged value is kUnknownHeapValue (same for a loop head with write side effects), we force the store to be kept in predecessors, since any load from a location with kUnknownHeapValue wont' be eliminated.

We don't build phis for merging heap values, so it sort of like forcing a spill to heap to do merge. Both store and load stay for a merge (think of loop head thing as a merge also, but handled a little differently).

Patch-set: 4
Label: Verified=0