Update patch set 1
Patch Set 1:

All of this input is coming from the OAT file. If we ensure that dex2oat only outputs valid utf8 then we in fact don't need to worry here.

Is it possible for a 3rd party to insert a malicious OAT file, or can they come only from a trusted dex2oat?

Patch-set: 1